Fixed bottom navbar from rebuilding
